+++ apps/server/src/api/routers/__init__.py
@@ -1,0 +1,13 @@
+from .diagram import router as diagram_router
+from .apikeys import router as apikeys_router
+from .files import router as files_router
+from .conversations import router as conversations_router
+from .websocket import router as websocket_router
+
+__all__ = [
+    "diagram_router",
+    "apikeys_router", 
+    "files_router",
+    "conversations_router",
+    "websocket_router"
+]
